chore: create useUriNavigate hook
