+++ src/containers/Breadcrumbs/Breadcrumbs.jsx
@@ -1,23 +1,13 @@
-import axios from 'axios'
 import { useEffect, useState, useRef } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { useLocation, useNavigate } from 'react-router-dom'
+import { useSelector } from 'react-redux'
+import { useLocation } from 'react-router-dom'
 import { toast } from 'react-toastify'
 import { InputText } from '@ynput/ayon-react-components'
 import * as Styled from './Breadcrumbs.styled'
 import HeaderButton from '../header/HeaderButton'
 
-import {
-  setFocusedFolders,
-  setFocusedProducts,
-  setFocusedVersions,
-  setFocusedRepresentations,
-  setFocusedTasks,
-  setFocusedWorkfiles,
-  setUri,
-  setUriChanged,
-} from '/src/features/context'
 import { upperFirst } from 'lodash'
+import useUriNavigate from '/src/hooks/useUriNavigate'
 
 const uri2crumbs = (uri = '', pathname) => {
   // parse uri to path and query params
@@ -67,8 +57,6 @@ const uri2crumbs = (uri = '', pathname) => {
 
 const Breadcrumbs = () => {
   const location = useLocation()
-  const dispatch = useDispatch()
-  const navigate = useNavigate()
 
   const [localUri, setLocalUri] = useState('')
   const [editMode, setEditMode] = useState(false)
@@ -97,131 +85,21 @@ const Breadcrumbs = () => {
     }
   }, [editMode])
 
-  const focusEntities = (entities) => {
-    const focusedFolders = []
-    const focusedProducts = []
-    const focusedVersions = []
-    const focusedRepresentations = []
-    const focusedTasks = []
-    const focusedWorkfiles = []
+  // NAVIGATE TO URI HOOK
+  const navigateToUri = useUriNavigate()
 
-    const project = entities[0].projectName
-
-    // assert we current url starts with projects/<projectName>
-    // if not, redirect
-
-    const path = window.location.pathname
-    if (!path.startsWith(`/projects/${project}`)) {
-      navigate(`/projects/${project}/browser`)
-    }
-
-    for (const entity of entities) {
-      if (entity.folderId) focusedFolders.push(entity.folderId)
-      if (entity.productId) focusedProducts.push(entity.productId)
-      if (entity.versionId) focusedVersions.push(entity.versionId)
-      if (entity.representationId) focusedRepresentations.push(entity.representationId)
-      if (entity.taskId) focusedTasks.push(entity.taskId)
-      if (entity.workfileId) focusedWorkfiles.push(entity.workfileId)
-
-      if (entity.projectName !== project) {
-        toast.error('Entities must be from the same project')
-        continue
-      }
-    }
-
-    dispatch(setFocusedFolders(focusedFolders))
-    dispatch(setFocusedProducts(focusedProducts))
-    dispatch(setFocusedVersions(focusedVersions))
-    dispatch(setFocusedRepresentations(focusedRepresentations))
-    dispatch(setFocusedTasks(focusedTasks))
-    dispatch(setFocusedWorkfiles(focusedWorkfiles))
-  }
-
-  const goThere = (e) => {
+  const goThere = async (e) => {
     e?.preventDefault()
     setEditMode(false)
     // blur input
     inputRef.current.blur()
 
-    if (!localUri) return
-
-    if (['ayon', 'ayon+entity'].includes(localUri.split('://')[0])) {
-      axios
-        .post('/api/resolve', { uris: [localUri] })
-        .then((res) => {
-          if (!res.data.length) {
-            toast.error('Could not resolve uri')
-            return
-          }
-          const entities = res.data[0].entities
-          if (!entities.length) {
-            toast.error('No entities found')
-            return
-          }
-          focusEntities(entities)
-          setTimeout(() => {
-            dispatch(setUri(res.data[0].uri))
-          }, 100)
-        })
-        .catch((err) => {
-          toast.error(err)
-        })
-        .finally(() => {
-          setEditMode(false)
-        })
-    } else if (localUri.startsWith('ayon+settings')) {
-      setEditMode(false)
-
-      //split query params
-
-      const [baseUri, query] = localUri.split('://')[1].split('?')
-
-      // extract addon name and version from uri
-      // ayon+settings://<addonName>:<addonVersion>/<settingsPathIncludingMoreSlashes>
-
-      const [addonStr, ...settingsPath] = baseUri.split('/')
-      const [addonName, addonVersion] = addonStr.split(':')
-
-      // parse query params
-
-      const qp = {}
-      if (query) {
-        for (const param of query.split('&')) {
-          const [key, value] = param.split('=')
-          qp[key] = value
-        }
-      }
-
-      let targetUrl = ''
-
-      if ('project' in qp && 'site' in qp) {
-        targetUrl = `manageProjects/siteSettings?`
-        targetUrl += `project=${qp.project}&site=${qp.site}`
-        targetUrl += `&addonName=${addonName}&addonVersion=${addonVersion}`
-        targetUrl += `&settingsPath=${settingsPath.join('|')}`
-      } else if ('project' in qp) {
-        targetUrl = `manageProjects/projectSettings?`
-        targetUrl += `project=${qp.project}`
-        targetUrl += `&addonName=${addonName}&addonVersion=${addonVersion}`
-        targetUrl += `&settingsPath=${settingsPath.join('|')}`
-      } else if ('site' in qp) {
-        targetUrl = `settings/site?`
-        targetUrl += `site=${qp.site}`
-        targetUrl += `&addonName=${addonName}&addonVersion=${addonVersion}`
-        targetUrl += `&settingsPath=${settingsPath.join('|')}`
-      } else {
-        targetUrl = `settings/studio`
-        targetUrl += `?addonName=${addonName}&addonVersion=${addonVersion}`
-        targetUrl += `&settingsPath=${settingsPath.join('|')}`
-      }
-
-      navigate(targetUrl)
-      dispatch(setUri(localUri))
-      dispatch(setUriChanged())
-      setEditMode(false)
-    } else {
-      toast.error('Invalid uri')
+    try {
+      await navigateToUri(localUri)
+    } catch (error) {
+      console.error(error)
       setLocalUri(ctxUri)
+    } finally {
       setEditMode(false)
     }
   }
